+++ redwood/.env.example
@@ -14,3 +14,5 @@ PUSHER_APP_ID=YOUR_PUSHER_APP_ID
 PUSHER_KEY=YOUR_PUSHER_KEY
 PUSHER_SECRET=YOUR_PUSHER_SECRET
 PUSHER_CLUSTER=YOUR_PUSHER_CLUSTER
+
+STARKNET_NOTARY_PRIVATE_KEY=YOUR_NOTARY_PRIVATE_KEY
